Typo修正

+++ web/database/migrations/2020_04_22_143136_carts.php
@@ -18,7 +18,7 @@ class Carts extends Migration
             $table->increments('id');
             $table->unsignedBigInteger('user_id');
             $table->unsignedBigInteger('product_id');
-            $table->integer('quantity');
+            $table->unsignedBigInteger('quantity');
             $table->timestamps();
 
             $table->foreign('user_id')->references('id')->on('users');
